Added migrations

+++ VidlyNew/Models/Movie.cs
@@ -19,7 +19,8 @@ namespace Vidly.Models
         
         [Display(Name = "Number in Stock")]
         [Range(1,20)]
-        public int NumberInStock { get; set; }
+        public byte NumberInStock { get; set; }
+
 
         public Genre Genre { get; set; }
